% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdaw}
\alias{bdaw}
\title{Blood draw data}
\format{
A data frame with the following columns:
\describe{
  \item{SUBJ}{Subject identifier}
  \item{VNUM}{Visit number}
  \item{AUC}{Area under the curve}
  \item{MET}{Device identifier}
}
}
\usage{
bdaw
}
\description{
Plasma cortisol area under curve (AUC) was calculated from the trapezoidal rule over the 12-h period of the hourly blood draws. The subjects were required to repeat the process in five visits. The aim of the agreement study was to assess how well the plasma cortisol AUC from hourly measurements agreed with plasma cortisol AUC that was measured every two hours.
}
\keyword{datasets}
