% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bfat}
\alias{bfat}
\title{Body fat data}
\format{
A data frame with the following columns:
\describe{
  \item{SUBJECT}{Subject identifier}
  \item{VISITNO}{Visit number}
  \item{BF}{Percentage body fat}
  \item{MET}{Device identifier}
}
}
\usage{
bfat
}
\description{
Percentage body fat was estimated from skinfold calipers and DEXA on a cohort of 90 adolescent girls. Skinfold caliper and DEXA measurements were taken at ages 12.5, 13 and 13.5. The objective was to determine the amount of agreement between the skinfold caliper and DEXA measurements of percentage body fat.
}
\keyword{datasets}
