\name{bpres}
\alias{bpres}
\title{Blood preasure data}
\description{
Systolic and diastolic blood pressure was measured in a sample of 384 subjects using a handle mercury sphygmomanometer device and an automatic device.}
\usage{bpres

}
\format{
ID	Number of subject
SEXO	Sex of subject
EDAD	Age of subject
SIS	Systolic blood preasure
DIA	Diastolic blood preasure
NM	Number of visit
METODE	Number of observer}

