\name{ccc-internal}
\alias{ccc-internal}
\alias{ic.ccc}
\alias{bdiag}
\alias{phi}
\title{Internal myPackage function}
\description{Internal myPackage functions}
\usage{
ic.ccc(ccc,dev,S)
bdiag(x)
phi(X1,Y1,X2,Y2,Dmat,delta)

}
\details{
These are not to be called by the user
}
\keyword{internal}
