% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cccrm-package.R
\docType{package}
\name{cccrm-package}
\alias{cccrm}
\alias{cccrm-package}
\title{cccrm: Concordance Correlation Coefficient for Repeated (and Non-Repeated) Measures}
\description{
Estimates the Concordance Correlation Coefficient to assess agreement. The scenarios considered are non-repeated measures, non-longitudinal repeated measures (replicates) and longitudinal repeated measures. It also includes the estimation of the one-way intraclass correlation coefficient also known as reliability index. The estimation approaches implemented are variance components and U-statistics approaches. Description of methods can be found in Fleiss (1986) \doi{10.1002/9781118032923} and Carrasco et al. (2013) \doi{10.1016/j.cmpb.2012.09.002}.
}
\author{
\strong{Maintainer}: Josep Lluis Carrasco \email{jlcarrasco@ub.edu}

Authors:
\itemize{
  \item Gonzalo Peon Pena
}

}
\keyword{internal}
