\name{cccrm}
\alias{cccrm}
 \title{CCC-RM}
\description{Concordance Correlation Coefficient for repeated measures}
\details{
Package:cccrm 

Title: CCC-RM

Version: 1.0

Date: 2011-03-01

Author: Josep Lluis Carrasco <jlcarrasco@ub.edu>, Josep Puig Martinez <j.puig04@gmail.com>

Maintainer: Josep Lluis Carrasco <jlcarrasco@ub.edu>

Depends: R (>= 2.08.0), nlme, gdata

Description:Concordance Correlation Coefficient for repeated measures

License: GPL version 2 or newer
}
\value{Functions:
\item{cccvc}{Concordance Correlation Coefficient by Variance Components}
\item{ccclon}{Longitudinal Repeated Measures Concordance Correlation Coefficient estimated by Variance Components}
\item{ccclonw}{Weighted Longitudinal Repeated Measures Concordance Correlation Coefficient estimated by Variance Components}
\item{cccUst}{Repeated Measures Concordance Correlation Coefficient estimated by U statistics}

}
