% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_boot_ci.R
\name{para_boot_ci}
\alias{para_boot_ci}
\title{Parametric bootstrap to make inference on the concordance correlation coefficient}
\usage{
para_boot_ci(
  dataset,
  model,
  ry,
  rind,
  rtime = NULL,
  rmet = NULL,
  vecD = NULL,
  covar = NULL,
  rho = 0,
  int = F,
  cl = 0.95,
  control.lme = list(),
  nboot = 500,
  parallel = TRUE,
  future_seed = 121
)
}
\arguments{
\item{dataset}{an object of class \code{data.frame}}

\item{model}{an object of class \code{lme}}

\item{ry}{Character string. Name of the outcome in the data set.}

\item{rind}{Character string. Name of the subject variable in the data set.}

\item{rtime}{Character string. Name of the time variable in the data set.}

\item{rmet}{Character string. Name of the method variable in the data set.}

\item{vecD}{Vector of weights. The length of the vector must be the same as the number of repeated measures.}

\item{covar}{Character vector. Name of covariates to include in the linear mixed model as fixed effects.}

\item{rho}{Within subject correlation structure. A value of 0 (default option) stands for compound symmetry and 1 is used for autorregressive of order 1 structure.}

\item{int}{Binary indicating if the subject-method interaction has to be included in the model when analyzing the non-longitudinal setting (defaults to FALSE).}

\item{cl}{Confidence level.}

\item{control.lme}{A list of control values for the estimation algorithm used in \code{lme} function. For further details see \code{lme} help.}

\item{nboot}{Number of bootstrap resamples.}

\item{parallel}{Whether the code is parallellized. The parallellization method is \code{multisession}.}

\item{future_seed}{Logical/Integer. The seed to be used for parallellization. Further details in \code{furrr_options}.}
}
\value{
Vector of CCC bootstrap estimates.
}
\description{
Parametric bootstrap to make inference on the concordance correlation coefficient
}
\keyword{internal}
