\name{bdaw}
\alias{bdaw}
\title{Blood draw data}
\description{Data in wich plasma cortisol AUC was calculated from the trapezoidal rule over the 12-hr period of the hourly blood draws. Data from five visits on each of subjects.}
\usage{bdaw

}
\format{
SUBJ	Number of subject

VNUM	Number of time

cort_suc	Plasma cortisol AUC

met	Number of observer
}

