\name{bfat}
\alias{bfat}
\title{Body fat data}
\description{Data in which percentage body fat was estimated from skinfold calipers and DEXA on a cohort of 90 adolescent girls. Skinfold caliper and DEXA measurements were taken at ages 12.5, 13 and 13.5.}
\usage{bfat

}
\format{
SUBJECT	Number of subject

VISITNO	Number of time

bf	Percentage bodi fat

met	Number of observer

}

