% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpres}
\alias{bpres}
\title{Blood pressure data}
\format{
A data frame with the following columns:
\describe{
  \item{ID}{Subject identifier}
  \item{SIS}{Systolic blood pressure in mmHg}
  \item{DIA}{Diastolic blood pressure in mmHg}
  \item{METODE}{Device identifier}
  \item{NM}{Identifier of replicates}
  \item{ALTURA}{Height in cm}
  \item{EDAD}{Age in years}
  \item{FRECUENC}{Heart rate}
  \item{INFOR_AR}{Have the subject been informed about he is hypertense?}
  \item{PESO}{Weight in Kg}
  \item{SEXO}{Gender. 1 for Male. 2 for Female}
  \item{TA}{Was the subject's blood pressure measured the last year? 1=Yes, 2=No, 9=Unknown}
  \item{TNSI_MED}{Does the subject receive treatment for hypertension? 1=Yes, 2=No, 3=Doubtful, 8=not applicable, 9=Insufficient data.}
}
}
\usage{
bpres
}
\description{
Systolic and diastolic blood pressure was measured in a sample of 384 subjects using a handle mercury sphygmomanometer device and an automatic device. The blood pressure was simultaneously measured twice by each instrument, thus every subject had four measurements, two by each method.
}
\keyword{datasets}
