% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ccclonw}
\alias{ccclonw}
\title{Concordance Correlation Coefficient estimated by Variance Components}
\usage{
ccclonw(
  dataset,
  ry,
  rind,
  rtime,
  rmet,
  vecD,
  covar = NULL,
  rho = 0,
  cl = 0.95,
  control.lme = list()
)
}
\arguments{
\item{dataset}{an object of class \code{data.frame}.}

\item{ry}{Character string. Name of the outcome in the data set.}

\item{rind}{Character string. Name of the subject variable in the data set.}

\item{rtime}{Character string. Name of the time variable in the data set.}

\item{rmet}{Character string. Name of the method variable in the data set.}

\item{vecD}{Vector of weights. The length of the vector must be the same as the number of repeated measures.}

\item{covar}{Character vector. Name of covariates to include in the linear mixed model as fixed effects.}

\item{rho}{Within subject correlation structure. A value of 0 (default option) stands for compound symmetry and 1 is used for autorregressive of order 1 structure.}

\item{cl}{Confidence level.}

\item{control.lme}{A list of control values for the estimation algorithm used in \code{lme} function. For further details see \code{lme} help.}
}
\description{
Concordance Correlation Coefficient for longitudinal repeated measures estimated by variance components
}
\details{
This function has been deprecated. See \link{ccc_vc}.
}
