% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{cccvc}
\alias{cccvc}
\title{Concordance Correlation Coefficient estimated by Variance Components}
\usage{
cccvc(
  dataset,
  ry,
  rind,
  rmet,
  covar = NULL,
  int = FALSE,
  cl = 0.95,
  control.lme = list()
)
}
\arguments{
\item{dataset}{an object of class \code{data.frame}.}

\item{ry}{Character string. Name of the outcome in the data set.}

\item{rind}{Character string. Name of the subject variable in the data set.}

\item{rmet}{Character string. Name of the method variable in the data set.}

\item{covar}{Character vector. Name of covariates to include in the linear mixed model as fixed effects.}

\item{int}{Binary indicating if the subject-method interaction has to be included in the model when analyzing the non-longitudinal setting (defaults to FALSE).}

\item{cl}{Confidence level.}

\item{control.lme}{A list of control values for the estimation algorithm used in \code{lme} function. For further details see \code{lme} help.}
}
\description{
Estimation of the concordance correlation coefficient for non-repeated measurements and non-longitudinal repeated measurements (replicates) using the variance components from a linear mixed model. The appropriate intraclass correlation coefficient is used as estimator of the concordance correlation coefficient.
}
\details{
This function has been deprecated. See \link{ccc_vc}.
}
